Refactoring controls to follow the documentation.
