+++ .gitignore
@@ -8,3 +8,5 @@ msr/__pycache__/routes.cpython-36.pyc
 msr/__pycache__/dao.cpython-36.pyc
 msr/__pycache__/manipula_arquivos.cpython-36.pyc
 msr/__pycache__/handler_threads.cpython-36.pyc
+msr/__pycache__/handler_files.cpython-36.pyc
+msr/__pycache__/authentication.cpython-36.pyc
