+++ msr/__init__.py
@@ -17,4 +17,5 @@ login_manager = LoginManager(app)
 login_manager.login_view = "login_page"
 login_manager.login_message_category = "info"
 
-from msr import routes
+from msr import main
+from msr import authentication
